requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
joblib
pyserial
jupyterlab